+++ components/Container.js
@@ -1,9 +1,10 @@
 import containerStyles from '../styles/Container.module.scss';
 
-export default function Container({ className, children }) {
+export default function Container({ className, children, styles }) {
   return (
     <div
       className={`${containerStyles.container} ${className ? className : ''}`}
+      style={styles}
     >
       {children}
     </div>