+++ CHANGELOG.md
@@ -28,3 +28,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - fixed scroll-x added by the layout container, adjust README
 - updated TwoColumn.js component with style props
 - updated favicon icon and add OG
+- normalized container use on components